requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy
opencv-python
decord
ffmpeg-python
moviepy
fastapi
uvicorn
python-multipart
runpod
einops
diffusers
transformers
accelerate
scikit-image
imageio
imageio-ffmpeg
pydantic
requests
